requirements.txt
djangorestframework
psycopg2-binary
environs